requirements.txt
numpy>=1.24
matplotlib>=3.7
gradio>=5.0
Pillow>=10.0
